+++ src/pint/models/pulsar_binary.py
@@ -3,19 +3,17 @@
 This module if for wrapping standalone binary models so that they work
 as PINT timing models.
 """
-import logging
 
 import astropy.units as u
 import numpy as np
 from astropy.time import Time
+from loguru import logger as log
 
 from pint import ls
 from pint.models.parameter import MJDParameter, floatParameter, prefixParameter
 from pint.models.stand_alone_psr_binaries import binary_orbits as bo
 from pint.models.timing_model import DelayComponent, MissingParameter, UnknownParameter
 from pint.utils import taylor_horner_deriv
-
-log = logging.getLogger(__name__)
 
 
 class PulsarBinary(DelayComponent):